edits to make it work when making new listings. Still needs a new_variety_page so we can make a list of varieties to start with

+++ src/main/java/com/launchcode/violetSwap/controllers/UserController.java
@@ -47,7 +47,7 @@ public class UserController {
             return "user/newUser";
         }
         userRepository.save(newUser);
-        int newUserId = newUser.getId();
+        long newUserId = newUser.getId();
         return "redirect:/user/"+newUserId;
     }
 }
